source/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
boto3==1.26.42
botocore==1.29.42
cachetools==5.2.0
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
fonttools==4.38.0
frozenlist==1.3.3
future==0.18.2
google-auth==2.15.0
google-auth-oauthlib==0.8.0
gspread==5.7.2
httplib2==0.21.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.4
line-bot-sdk==2.3.0
MarkupSafe==2.1.1
matplotlib==3.6.2
multidict==6.0.4
numpy==1.24.1
oauth2client==4.1.3
oauthlib==3.2.2
packaging==22.0
pandas==1.5.2
Pillow==9.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
seaborn==0.12.2
six==1.16.0
urllib3==1.26.13
Werkzeug==2.2.2
yarl==1.8.2
zipp==3.11.0
